refactor in all app

+++ lib/src/feature/salary_details/logic/model/salary_details_model.dart
@@ -46,7 +46,7 @@ class SalaryDetailsModel {
 }
 
 class Commissions {
-  int? amount;
+  double? amount;
   String? description;
   String? date;
